Don't use planes currently in-use on other CRTCs on first commit
When doing the first commit, libliftoff will try to disable all
candidate planes. This is necessary to make sure we start atomic
test commits from a clean state, and are able to incrementally come
up with a combination which works.

However, disabling a plane currently in-use on another CRTC (e.g.
a primary plane) will result in failed commits. We can't steal that
plane without removing it from the other CRTC first.

Add a new liftoff_plane.crtc_id field populated with the current
CRTC ID for the plane.